+++ packages/frontend/src/presenter/DataSources/useThriftHandler.ts
@@ -46,7 +46,7 @@ export const useThriftHandler = ({ state, updateState, step }: HandlerStep) => {
     schemaHeaders: objectToFieldArray(schemaHeaders),
     operationHeaders: objectToFieldArray(operationHeaders),
   });
-  const { handleSubmit, errors, control } = useForm({
+  const { handleSubmit, errors, control, register } = useForm({
     resolver: vestResolver(suite),
     reValidateMode: 'onSubmit',
     defaultValues,
@@ -94,6 +94,7 @@ export const useThriftHandler = ({ state, updateState, step }: HandlerStep) => {
     onSubmit: handleSubmit(onSubmit),
     errors,
     control,
+    register,
     schemaFields,
     appendSchemaField,
     removeSchemaField,